BE/Feat/#172 : ãqr코드 추가 및 토큰 수정

+++ Backend/src/main/java/ssafy/d210/backend/service/PaymentService.java
@@ -4,7 +4,9 @@ import org.springframework.stereotype.Service;
 import ssafy.d210.backend.dto.common.ResponseSuccessDto;
 import ssafy.d210.backend.dto.request.transaction.SignedRequest;
 
+import java.math.BigInteger;
+
 
 public interface PaymentService {
-    public ResponseSuccessDto<Boolean> increaseToken(long userId, int quantity);
+    public ResponseSuccessDto<Boolean> increaseToken(long userId, BigInteger quantity);
 }
